6.0

+++ app/src/main/cpp/mp4v2-2.0.0/libutil/Utility.cpp
@@ -531,26 +531,26 @@ Utility::process_impl()
                 printHelp( false, false );
                 return SUCCESS;
 
-            case LC_DEBUG:
+            case (int)LC_DEBUG:
                 debugUpdate( std::strtoul( prog::optarg, NULL, 0 ) );
                 break;
 
-            case LC_VERBOSE:
+            case (int)LC_VERBOSE:
             {
                 const uint32_t level = std::strtoul( prog::optarg, NULL, 0 );
                 _verbosity = ( level < 4 ) ? level : 3;
                 break;
             }
 
-            case LC_HELP:
+            case (int)LC_HELP:
                 printHelp( true, false );
                 return SUCCESS;
 
-            case LC_VERSION:
+            case (int)LC_VERSION:
                 printVersion( false );
                 return SUCCESS;
 
-            case LC_VERSIONX:
+            case (int)LC_VERSIONX:
                 printVersion( true );
                 return SUCCESS;
